Delayed processing for ProcessManager.pidToProcessInfo (#321)
Uses TraceProcessor.ProcessedUntil mechanism to guarantee that process metadata
is not discarded before all relevant trace events have been processed.
